feat: implement JWT authentication and secure expense ownership

+++ ExpenseTracker.API/Expenses/Service/ExpenseService.cs
@@ -17,29 +17,35 @@ public class ExpenseService : IExpenseService
         var expenses = await _repository.GetAllAsync();
         return expenses.Select(MapToDto).ToList();
     }
-
-    public async Task<ExpenseResponseDto?> GetByIdAsync(int id)
+    
+    public async Task<List<ExpenseResponseDto>> GetAllByUserId(int userId)
     {
-        var expense = await _repository.GetByIdAsync(id);
-        return expense is null ? null : MapToDto(expense);
+        var expenses = await _repository.GetExpesnesByUserId(userId);
+        return expenses.Select(MapToDto).ToList();
     }
 
-    public async Task<ExpenseResponseDto> Create(ExpenseCreateDto expense)
+    public async Task<ExpenseResponseDto?> GetByIdAsync(int id, int userId)
+    {
+        var expense = await _repository.GetByIdAsync(id);
+        if (expense == null || expense.UserId != userId) return null;
+        return MapToDto(expense);
+    }
+
+    public async Task<ExpenseResponseDto> Create(ExpenseCreateDto expense, int userId)
     {
         var createdExpense = new Expense(
             expense.ExpenseName,
             expense.Amount,
             expense.Date,
-            expense.UserId);
+            userId);
         var created = await _repository.CreateAsync(createdExpense);
         return MapToDto(created);
     }
 
-    public async Task<ExpenseResponseDto?> Update(ExpenseUpdateDto updatedExpense, int id)
+    public async Task<ExpenseResponseDto?> Update(ExpenseUpdateDto updatedExpense, int id, int userId)
     {
         var expense = await _repository.GetByIdAsync(id);
-        if (expense is null)
-            return null;
+        if (expense is null || expense.UserId != userId) return null;
 
         expense.ExpenseName = updatedExpense.ExpenseName;
         expense.Amount = updatedExpense.Amount;
@@ -50,12 +56,14 @@ public class ExpenseService : IExpenseService
         return MapToDto(expense);
     }
 
-    public async Task<bool> Delete(int id)
+    public async Task<bool> Delete(int id, int userId)
     {
+        var expense = await _repository.GetByIdAsync(id);
+        if (expense is null || expense.UserId != userId) return false;
+        
         return await _repository.DeleteAsync(id);
     }
     
-    // Mapper for now
     private ExpenseResponseDto MapToDto(Expense expense)
     {
         return new ExpenseResponseDto(
